[TLC-249] Refactor, include controller 'find DSO' operation
