+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/repository/IdentifierRestRepository.java
@@ -7,25 +7,31 @@
  */
 package org.dspace.app.rest.repository;
 
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+
+import java.io.IOException;
+import java.net.URI;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Iterator;
+import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
-import java.util.function.Function;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.NotSupportedException;
 
+import org.atteo.evo.inflector.English;
 import org.dspace.app.rest.DiscoverableEndpointsService;
-import org.dspace.app.rest.IdentifierRestController;
 import org.dspace.app.rest.Parameter;
 import org.dspace.app.rest.SearchRestMethod;
 import org.dspace.app.rest.exception.LinkNotFoundException;
 import org.dspace.app.rest.exception.RESTAuthorizationException;
 import org.dspace.app.rest.exception.RepositoryMethodNotImplementedException;
 import org.dspace.app.rest.exception.UnprocessableEntityException;
+import org.dspace.app.rest.model.DSpaceObjectRest;
 import org.dspace.app.rest.model.IdentifierRest;
 import org.dspace.app.rest.repository.handler.service.UriListHandlerService;
+import org.dspace.app.rest.utils.ContextUtil;
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.content.DSpaceObject;
 import org.dspace.content.Item;
@@ -36,24 +42,39 @@ import org.dspace.handle.service.HandleService;
 import org.dspace.identifier.DOI;
 import org.dspace.identifier.DOIIdentifierProvider;
 import org.dspace.identifier.IdentifierException;
+import org.dspace.identifier.IdentifierNotFoundException;
+import org.dspace.identifier.IdentifierNotResolvableException;
+import org.dspace.identifier.factory.IdentifierServiceFactory;
 import org.dspace.identifier.service.DOIService;
 import org.dspace.identifier.service.IdentifierService;
 import org.dspace.services.factory.DSpaceServicesFactory;
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
+import org.springframework.hateoas.Link;
+import org.springframework.hateoas.TemplateVariable;
+import org.springframework.hateoas.TemplateVariables;
+import org.springframework.hateoas.UriTemplate;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
- * Controller for exposition of vocabularies entry details for the submission
+ * Item REST Repository and Controller for persistent identifiers.
+ * The controller annotation and endpoint registration allows the "find DSO by identifier" method which was
+ * previously implmented in org.dspace.app.rest.IdentifierRestController
  *
- * @author Andrea Bollini (andrea.bollini at 4science.it)
+ * @author Kim Shepherd
  */
-@Component("pid.identifier")
-public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRest, String> {
+@RestController
+@RequestMapping("/api/" + IdentifierRestRepository.CATEGORY)
+@Component(IdentifierRestRepository.CATEGORY + "." + IdentifierRestRepository.NAME)
+public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRest, String> implements InitializingBean {
     @Autowired
     private DiscoverableEndpointsService discoverableEndpointsService;
     @Autowired
@@ -67,6 +88,26 @@ public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRes
     @Autowired
     private IdentifierService identifierService;
 
+    // Set category and name for routing
+    public static final String CATEGORY = "pid";
+    public static final String NAME = "identifier";
+
+    /**
+     * Register /api/pid/find?id=... as a discoverable endpoint service
+     *
+     * @throws Exception
+     */
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        discoverableEndpointsService
+                .register(this,
+                        Arrays.asList(Link.of(UriTemplate.of("/api/pid/find",
+                                                new TemplateVariables(
+                                                        new TemplateVariable("id",
+                                                                TemplateVariable.VariableType.REQUEST_PARAM))),
+                                        CATEGORY)));
+    }
+
     @PreAuthorize("permitAll()")
     @Override
     public Page<IdentifierRest> findAll(Context context, Pageable pageable) {
@@ -75,7 +116,16 @@ public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRes
         return new PageImpl<>(results, pageable, 0);
     }
 
+    /**
+     * Find the identifier object for a given identifier string (eg. doi)
+     * @param context
+     *            the dspace context
+     * @param identifier
+     *            the rest object id
+     * @return
+     */
     @PreAuthorize("permitAll()")
+    @RequestMapping(method = RequestMethod.GET)
     @Override
     public IdentifierRest findOne(Context context, String identifier) {
         DSpaceObject dso;
@@ -105,14 +155,20 @@ public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRes
                     identifierRest.setValue(handleService.getCanonicalForm(dso.getHandle()));
                 }
             } else {
-                throw new LinkNotFoundException(IdentifierRestController.CATEGORY, IdentifierRest.NAME, identifier);
+                throw new LinkNotFoundException(IdentifierRestRepository.CATEGORY, IdentifierRest.NAME, identifier);
             }
         } catch (SQLException | IdentifierException e) {
-            throw new LinkNotFoundException(IdentifierRestController.CATEGORY, IdentifierRest.NAME, identifier);
+            throw new LinkNotFoundException(IdentifierRestRepository.CATEGORY, IdentifierRest.NAME, identifier);
         }
         return identifierRest;
     }
 
+    /**
+     * Find identifiers associated with a given item
+     * @param uuid
+     * @param pageable
+     * @return
+     */
     @SearchRestMethod(name = "findByItem")
     @PreAuthorize("permitAll()")
     public Page<IdentifierRest> findByItem(@Parameter(value = "uuid", required = true)
@@ -132,19 +188,20 @@ public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRes
                 results.add(new IdentifierRest(handleUrl, "handle", null));
             }
         } catch (SQLException | IdentifierException e) {
-            throw new LinkNotFoundException(IdentifierRestController.CATEGORY, IdentifierRest.NAME, uuid);
+            throw new LinkNotFoundException(IdentifierRestRepository.CATEGORY, IdentifierRest.NAME, uuid);
         }
         // Return list of identifiers for this DSpaceObject
         return new PageImpl<>(results, pageable, results.size());
     }
 
     /**
-     * Right now, the only supported identifier type is DOI
+     * Create (mint / queue for registration) a new persistent identifier of a given type (eg DOI) for an item
+     * Currently, the only supported identifier type for this operation is "doi"
+     *
      * @param context
      *            the dspace context
      * @param list
-     *            The list of Strings that will be used as data for the object that's to be created
-     *            This list is retrieved from the uri-list body
+     *            A uri-list with the item URI for which to create an identifier
      * @return
      * @throws AuthorizeException
      * @throws SQLException
@@ -187,6 +244,15 @@ public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRes
         return identifierRest;
     }
 
+    /**
+     * Perform DOI registration, skipping any other filters used.
+     *
+     * @param context
+     * @param item
+     * @return
+     * @throws SQLException
+     * @throws AuthorizeException
+     */
     private IdentifierRest registerDOI(Context context, Item item)
             throws SQLException, AuthorizeException {
         String identifier = null;
@@ -214,6 +280,51 @@ public class IdentifierRestRepository extends DSpaceRestRepository<IdentifierRes
         itemService.updateLastModified(context, item);
         context.complete();
         return identifierRest;
+    }
+
+
+    /**
+     * Redirect to a DSO page, given an identifier
+     *
+     * @param request   HTTP request
+     * @param response  HTTP response
+     * @param id        The persistent identifier (eg. handle, DOI) to search for
+     * @throws IOException
+     * @throws SQLException
+     */
+    @RequestMapping(method = RequestMethod.GET, value = "find", params = "id")
+    @SuppressWarnings("unchecked")
+    public void getDSObyIdentifier(HttpServletRequest request,
+                                   HttpServletResponse response,
+                                   @RequestParam("id") String id)
+            throws IOException, SQLException {
+
+        DSpaceObject dso;
+        Context context = ContextUtil.obtainContext(request);
+        IdentifierService identifierService = IdentifierServiceFactory
+                .getInstance().getIdentifierService();
+        try {
+            // Resolve identifier to a DSpace object
+            dso = identifierService.resolve(context, id);
+            if (dso != null) {
+                // Convert and respond with a redirect to the object itself
+                DSpaceObjectRest dsor = converter.toRest(dso, utils.obtainProjection());
+                URI link = linkTo(dsor.getController(), dsor.getCategory(),
+                        English.plural(dsor.getType()))
+                        .slash(dsor.getId()).toUri();
+                response.setStatus(HttpServletResponse.SC_FOUND);
+                response.sendRedirect(link.toString());
+            } else {
+                // No object could be found
+                response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+            }
+        } catch (IdentifierNotFoundException e) {
+            response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+        } catch (IdentifierNotResolvableException e) {
+            response.setStatus(HttpServletResponse.SC_NOT_IMPLEMENTED);
+        } finally {
+            context.abort();
+        }
     }
 
     @Override
